fix(dts-generator): fix download-apijson by querying differently (#489)
Don't use the scoped search, which apparently stopped working, and use standard search instead.

The max supported limit of search results is 250 right now
and the current number of hits is still lower (183).
This number might increase over time to a degree
where relevant libraries are not within the first 250.
Results are sorted by popularity etc., so the libraries are right now still
at the very beginning of the result list.
This creates a comfortable buffer for the time being.

Fixes https://github.com/SAP/ui5-typescript/issues/488